libhost1x: remove nvhost-display.h nvhost-gr2d.h nvhost-gr3d.h
Move theirs stuff to nvhost.h, it's quite small.
