--- src/libhost1x/nvhost-display.h
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2017 Dmitry Osipenko <digetx@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
-#ifndef GRATE_HOST1X_NVHOST_DISP_H
-#define GRATE_HOST1X_NVHOST_DISP_H
-
-#include <linux/types.h>
-
-#include "host1x.h"
-#include "nvhost.h"
-
-struct nvhost_display {
-	struct host1x_display base;
-	struct host1x_overlay overlay;
-	struct nvhost *nvhost;
-	uint32_t vblank_syncpt;
-	int plane;
-	int fd;
-};
-
-struct nvhost_display * nvhost_display_create(struct nvhost *nvhost);
-void nvhost_display_close(struct nvhost_display *display);
-
-#endif
